+++ src/main/java/roomescape/config/LoginMember.java
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface AuthenticationPrincipal {
+public @interface LoginMember {
 }
